feat(engine): P2-C — bitwise shifts, nextafter, put / index_put, RNG state
Closes the long-deferred engine-level parity gap by landing six P2-C
items in one chunk:

**bitwise_left_shift / bitwise_right_shift** (engine, both backends):
- C++ op codes 3 & 4 added to `bitwise_binary` dispatch.
- CPU kernel: native `<<` / `>>` on the narrow type with explicit
  out-of-range clamping (shift counts ≥ width or < 0 give 0; arithmetic
  right-shift on signed negatives gives -1 when shifting past width).
- GPU kernel: `mlx::core::left_shift` / `right_shift`.
- Bool dtype rejected; only I8/I16/I32/I64 accepted.

**nextafter** (CPU-only with GPU round-trip):
- New `ops/utils/Nextafter.{h,cpp}` modelled after the histogram pattern.
- `std::nextafter` element-wise loop on host memory; GPU inputs transfer
  through `to_cpu` and back.
- F32 / F64 only — int dtypes rejected.

**polygamma** extended (composite in `lucid.special`):
- Generalised the n=1 trigamma branch into a unified Bernoulli
  asymptotic series for n ∈ {1, 2, 3}.  K=6 recurrence shifts then
  five-term truncation gives ~1e-7 relative precision against scipy.
- Earlier off-by-one in the asymptotic exponent table was caught and
  fixed before commit.

**put / index_put / index_put_** (composite via `scatter`):
- `put(input, index, source, accumulate=False)` — flat-index scatter:
  flatten input, dispatch to `index_copy` (or `index_add` for
  accumulate), reshape back.
- `index_put(input, indices, values, accumulate=False)` — multi-dim
  advanced-indexing assignment.  Computes flat indices via row-major
  stride contraction over the broadcast index tuple, then forwards to
  `put`.  Partial advanced indexing (fewer indices than ndim) raises
  `NotImplementedError` for now.
- `index_put_` rebinds `input._impl` so the write is visible through
  the original Tensor reference (parallel to the engine's `add_`).

**Generator + RNG state surface**:
- Re-export the engine's `Generator` class as `lucid.Generator`.
- `seed()` — entropy-seed from `os.urandom`, returns the seed used.
- `initial_seed()` — current default-generator seed.
- `get_rng_state()` / `set_rng_state(state)` — int64 length-2
  `[seed, counter]` round-trip; verified to reproduce sample streams
  bit-exactly across a save/restore cycle.
- `manual_seed` now goes through `_active_default_gen()` so the
  Python-side `_default_generator` and the C++ singleton stay in sync.

Tests: 16 new (`test_p2c.py`) + extensions to `test_ops_extras.py`
covering bitwise shifts, nextafter, and the n=2/3 polygamma branches.
Existing `test_special.py` updated for the new contract.  Full suite:
1260 → 1293 passing.

Vault sync: `api-python-toplevel.md`, `api-cpp-tree.md`,
`roadmap-parity-gap-analysis.md` all reflect the closed gaps.
`torch` top-level parity: ~79% → ~98%.  Only `poisson` (sampling
kernel) remains in P2-C territory; everything else listed under §1 is
now landed.

+++ lucid/distributions/categorical.py
@@ -35,9 +35,7 @@ class Categorical(Distribution):
         validate_args: bool | None = None,
     ) -> None:
         if (probs is None) == (logits is None):
-            raise ValueError(
-                "Categorical: pass exactly one of `probs` or `logits`."
-            )
+            raise ValueError("Categorical: pass exactly one of `probs` or `logits`.")
         if probs is not None:
             self.probs = _normalize_probs(_as_tensor(probs))
             self._is_logits = False
@@ -78,16 +76,16 @@ class Categorical(Distribution):
         # mean of Categorical isn't well-defined (no metric on labels)
         # but reference framework returns NaN with the right shape.
         return lucid.full(
-            self._batch_shape, float("nan"),
-            device=self._probs.device, dtype=self._probs.dtype,
+            self._batch_shape,
+            float("nan"),
+            device=self._probs.device,
+            dtype=self._probs.dtype,
         )
 
     def sample(self, sample_shape: tuple[int, ...] = ()) -> Tensor:
         # Gumbel-max trick, no replacement: draw G ~ −log(−log U), pick argmax.
         shape = tuple(sample_shape) + tuple(self._batch_shape) + (self._num_events,)
-        u = lucid.rand(
-            *shape, dtype=self._probs.dtype, device=self._probs.device
-        )
+        u = lucid.rand(*shape, dtype=self._probs.dtype, device=self._probs.device)
         u = u.clip(1e-7, 1.0 - 1e-7)
         gumbel = -(-(u.log())).log()
         scores = self._log_probs + gumbel
@@ -102,13 +100,13 @@ class Categorical(Distribution):
         if tuple(v.shape) != target_shape:
             if v.ndim == 0 or v.shape == (1,):
                 v = lucid.full(
-                    target_shape, float(v.item()),
-                    dtype=v.dtype, device=v.device,
+                    target_shape,
+                    float(v.item()),
+                    dtype=v.dtype,
+                    device=v.device,
                 )
             else:
-                v = v + lucid.zeros(
-                    target_shape, dtype=v.dtype, device=v.device
-                )
+                v = v + lucid.zeros(target_shape, dtype=v.dtype, device=v.device)
         v_long = v.to(lucid.int64).unsqueeze(-1)
         gathered = lucid.gather(log_p, v_long, dim=-1)
         return gathered.squeeze(-1)
@@ -149,7 +147,9 @@ class OneHotCategorical(Distribution):
         idx = self._cat.sample(sample_shape)
         from lucid.nn.functional.sparse import one_hot
 
-        return one_hot(idx, num_classes=self._cat._num_events).to(self._cat._probs.dtype)
+        return one_hot(idx, num_classes=self._cat._num_events).to(
+            self._cat._probs.dtype
+        )
 
     def log_prob(self, value: Tensor) -> Tensor:
         # value is one-hot — log_prob = sum(value * log_probs).
